requirements.txt
Django==5.0.3
python-decouple==3.8
mysqlclient==2.1.1